+++ contrib/lua/CMakeLists.txt
@@ -2,7 +2,7 @@ project(lua LANGUAGES C)
 
 set(BUILD_SHARED_LIBS OFF)
 
-file(GLOB SRC_FILES *.c)
+file(GLOB SRC_FILES src/*.c)
 
 add_definitions(-DLUA_USE_APICHECK)
 if (NOT WIN32)
@@ -10,3 +10,4 @@ if (NOT WIN32)
 endif (NOT WIN32)
 
 add_library(${PROJECT_NAME} STATIC ${SRC_FILES})
+target_include_directories(${PROJECT_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src)
