+++ contrib/lua/lua.hpp
@@ -3,7 +3,7 @@
 // <<extern "C">> not supplied automatically because Lua also compiles as C++
 
 extern "C" {
-#include "lua.h"
-#include "lualib.h"
-#include "lauxlib.h"
+#include "src/lua.h"
+#include "src/lualib.h"
+#include "src/lauxlib.h"
 }
