+++ contrib/lua/src/lbitlib.c
