+++ contrib/lua/src/linit.c
