+++ contrib/lua/src/lstring.c
